finalize migrator implementation
