+++ src/common/database/migrator/index.ts
@@ -1,0 +1,1 @@
+export * from './migrator';
