+++ src/common/database/migrator/sample-migration.ts
@@ -1,0 +1,40 @@
+import sequelize, { DataTypes } from 'sequelize';
+import { IMigratorOptions } from '../type';
+
+export const up = async ({ context }: IMigratorOptions): Promise<void> => {
+  await context.createTable('examples', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      allowNull: false,
+      primaryKey: true,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    foreignId: {
+      type: DataTypes.UUID,
+      references: {
+        model: 'examples',
+        key: 'id',
+      },
+      onDelete: 'CASCADE',
+      onUpdate: 'CASCADE',
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: sequelize.literal('CURRENT_TIMESTAMP'),
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: sequelize.literal('CURRENT_TIMESTAMP'),
+    },
+  });
+};
+
+export const down = async ({ context }: IMigratorOptions): Promise<void> => {
+  await context.dropTable('examples');
+};
