+++ src/common/database/type/index.ts
@@ -1,0 +1,2 @@
+export * from './migrator.interface';
+export * from './migration.interface';
