+++ src/common/database/type/migration.interface.ts
@@ -1,0 +1,4 @@
+export interface IMigrationMeta {
+  name: string;
+  path?: string;
+}
